pkg/cover: split logic, move some data processing to pkg/coveragedb
